requirements.txt
asgiref==3.7.2distlib==0.3.7Django==4.2.5django-phonenumber-field==7.1.0filelock==3.12.3pymysql==1.0.2phonenumbers==8.13.1platformdirs==3.10.0secure-smtplib==0.1.1sqlparse==0.4.4typing-extensions==4.7.1tzdata==2023.3virtualenv==20.4.4